make RtfColorTable.getColorNumber return an Integer instead of int, so that it can return a null value -- submitted by Peter Herweg, see http://nagoya.apache.org/bugzilla/show_bug.cgi?id=23082
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@196915 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/rtf/renderer/RTFHandler.java
@@ -452,7 +452,7 @@ public class RTFHandler extends FOInputHandler {
            white. */
         if ((fopValue.getRed() == 0) && (fopValue.getGreen() == 0)
                 && (fopValue.getBlue() == 0) && (fopValue.getAlpha() == 0)) {
-            rtfColor = RtfColorTable.getInstance().getColorNumber("white");
+            rtfColor = RtfColorTable.getInstance().getColorNumber("white").intValue();
         } else {
             rtfColor = convertFOPColorToRTF(fopValue);
         }
@@ -472,7 +472,7 @@ public class RTFHandler extends FOInputHandler {
         int greenComponent = ColorType.convertChannelToInteger (fopColor.getGreen());
         int blueComponent = ColorType.convertChannelToInteger (fopColor.getBlue());
         return RtfColorTable.getInstance().getColorNumber(redComponent,
-                greenComponent, blueComponent);
+                greenComponent, blueComponent).intValue();
     }
 
 }
